+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Fixed a bug where search inputs within element selector modals weren’t getting focus rings.
 - Fixed a bug where boolean menu inputs were initially treating `null` values as `true`.
 - Fixed a bug where boolean menu inputs weren’t toggling other fields. ([#13613](https://github.com/craftcms/cms/issues/13613))
+- Fixed a bug where `Craft.namespaceId()` wasn’t working properly if the namespace ended in a `]` character.
 
 ## 4.5.2 - 2023-08-24